chore: move constant from common to server
