+++ packages/oraidex-common/src/pairs.ts
@@ -170,52 +170,3 @@ export const PAIRS_CHART = PAIRS.map((pair) => {
     info: `${assets[0]}-${assets[1]}`
   };
 });
-
-// The pairs are rearranged in the correct order of base and quote
-export const ARRANGED_PAIRS = PAIRS.map((pair) => {
-  const pairDenoms = pair.asset_infos.map((assetInfo) => parseAssetInfoOnlyDenom(assetInfo));
-  if (pairDenoms.some((denom) => denom === ORAI) && pairDenoms.some((denom) => denom === INJECTIVE_CONTRACT))
-    return {
-      ...pair,
-      asset_infos: [
-        ORAI_INFO,
-        {
-          token: {
-            contract_addr: INJECTIVE_CONTRACT
-          }
-        } as AssetInfo
-      ],
-      symbols: ["ORAI", "INJ"]
-    } as PairMapping;
-
-  if (
-    pairDenoms.some((denom) => denom === ORAIX_CONTRACT) &&
-    pairDenoms.some((denom) => denom === USDC_CONTRACT)
-  )
-    return {
-      ...pair,
-      asset_infos: [
-        {
-          token: {
-            contract_addr: ORAIX_CONTRACT
-          }
-        } as AssetInfo,
-        {
-          token: {
-            contract_addr: USDC_CONTRACT
-          }
-        } as AssetInfo
-      ],
-      symbols: ["ORAIX", "USDC"]
-    } as PairMapping;
-  return pair;
-});
-
-export const ARRANGED_PAIRS_CHART = ARRANGED_PAIRS.map((pair) => {
-  const assets = pair.asset_infos.map(parseAssetInfoOnlyDenom);
-  return {
-    ...pair,
-    symbol: `${pair.symbols[0]}/${pair.symbols[1]}`,
-    info: `${assets[0]}-${assets[1]}`
-  };
-});
